// src/app/shared/models/apptypes.model.ts
export class ApptypesDto {
    constructor() {
    }
    atId: string;
    atCode: string | null;
    atName: string | null;
    atDescription: string | null;
    atStatus: number | null;
    atCreateddate: Date | null;
    atCreatedby: string | null;
    atUpdateddate: Date | null;
    atUpdatedby: string | null;
  }